fix(helm): fixed helms payload should only print payload lines when flag is activated
