+++ pkg/kics/resolver_sink.go
@@ -43,15 +43,16 @@ func (s *Service) resolverSink(ctx context.Context, filename, scanID string) ([]
 			}
 
 			file := model.FileMetadata{
-				ID:           uuid.New().String(),
-				ScanID:       scanID,
-				Document:     document,
-				OriginalData: string(rfile.OriginalData),
-				Kind:         kind,
-				FilePath:     rfile.FileName,
-				Content:      string(rfile.Content),
-				HelmID:       rfile.SplitID,
-				IDInfo:       rfile.IDInfo,
+				ID:               uuid.New().String(),
+				ScanID:           scanID,
+				Document:         PrepareScanDocument(document, kind),
+				OriginalData:     string(rfile.OriginalData),
+				LineInfoDocument: document,
+				Kind:             kind,
+				FilePath:         rfile.FileName,
+				Content:          string(rfile.Content),
+				HelmID:           rfile.SplitID,
+				IDInfo:           rfile.IDInfo,
 			}
 			s.saveToFile(ctx, &file)
 		}
