+++ pkg/kics/sink.go
@@ -63,51 +63,6 @@ func (s *Service) sink(ctx context.Context, filename, scanID string, rc io.Reade
 	return errors.Wrap(err, "failed to save file content")
 }
 
-// PrepareScanDocument removes _kics_lines from payload and parses json filters
-func PrepareScanDocument(body map[string]interface{}, kind model.FileKind) map[string]interface{} {
-	bodyMap := make(map[string]interface{})
-	j, err := json.Marshal(body)
-	if err != nil {
-		log.Error().Msgf("failed to remove kics line information")
-		return body
-	}
-	if err := json.Unmarshal(j, &bodyMap); err != nil {
-		log.Error().Msgf("failed to remove kics line information")
-		return body
-	}
-	prepareScanDocumentRoot(bodyMap, kind)
-	return bodyMap
-}
-
-func prepareScanDocumentRoot(body interface{}, kind model.FileKind) {
-	switch bodyType := body.(type) {
-	case map[string]interface{}:
-		prepareScanDocumentValue(bodyType, kind)
-	case []interface{}:
-		for _, indx := range bodyType {
-			prepareScanDocumentRoot(indx, kind)
-		}
-	}
-}
-
-func prepareScanDocumentValue(bodyType map[string]interface{}, kind model.FileKind) {
-	delete(bodyType, "_kics_lines")
-	for key, v := range bodyType {
-		switch value := v.(type) {
-		case map[string]interface{}:
-			prepareScanDocumentRoot(value, kind)
-		case []interface{}:
-			for _, indx := range value {
-				prepareScanDocumentRoot(indx, kind)
-			}
-		case string:
-			if field, ok := lines[kind]; ok && contains(field, key) {
-				bodyType[key] = resolveJSONFilter(value)
-			}
-		}
-	}
-}
-
 func contains(strSlice []string, key string) bool {
 	for _, v := range strSlice {
 		if v == key {
